+++ lib/app/app.dart
@@ -1,0 +1,79 @@
+import 'package:caronas_usp/app/modules/pegar/ui/pegar_page.dart';
+import 'package:caronas_usp/app/modules/historico/ui/historico_page.dart';
+import 'package:caronas_usp/app/modules/oferecer/ui/oferecer_page.dart';
+import 'package:caronas_usp/app/modules/profile/ui/profile_page.dart';
+import 'package:flutter/material.dart';
+
+class App extends StatelessWidget {
+  const App({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+
+  int _selectedIndex = 0;
+
+  final screens = const [
+    PegarPage(),
+    HistoricoPage(),
+    OferecerPage(),
+    ProfilePage()
+  ];
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: IndexedStack(
+        index: _selectedIndex,
+        children: screens,
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        type: BottomNavigationBarType.fixed,
+        backgroundColor: Colors.green,
+        selectedItemColor: Colors.white,
+        unselectedItemColor: Colors.white38,
+        iconSize: 25,
+        showUnselectedLabels: false,
+        currentIndex: _selectedIndex,
+        onTap: (index) => setState(() => _selectedIndex = index) ,
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.thumb_up_rounded),
+            label: 'Pegar',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.people_rounded),
+            label: 'Caronas',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.directions_car_filled),
+            label: 'Oferecer',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.account_circle_rounded),
+            label: 'Perfil',
+          ),
+        ],
+      ),
+    );
+  }
+}
