+++ lib/app/modules/pegar/ui/pegar_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class PegarCaronas extends StatelessWidget {
-  const PegarCaronas({Key? key}) : super(key: key);
+class PegarPage extends StatelessWidget {
+  const PegarPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => Scaffold(
