+++ lib/app/modules/profile/ui/profile_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class Perfil extends StatelessWidget {
-  const Perfil({Key? key}) : super(key: key);
+class ProfilePage extends StatelessWidget {
+  const ProfilePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => Scaffold(
